+++ CHANGELOG.md
@@ -5,6 +5,10 @@ once the API is stable.
 
 ## Unreleased
 
+- ADDED: `message.make_privmsgs` now takes `third_person`.
+
+  This can be used to create third person messages. (`/me likes this :D`)
+
 - FIXED: `parsers.nick`.
 
   No longer falls over on nicks that are already just nicks (no ident, etc).